Move Avro generics to core.
