+++ core/src/main/java/org/apache/iceberg/data/avro/DataWriter.java
