+++ core/src/main/java/org/apache/iceberg/data/avro/GenericWriters.java
