+++ data/src/main/java/org/apache/iceberg/data/InternalRecordWrapper.java
@@ -29,6 +29,7 @@ import java.util.function.Function;
 import org.apache.iceberg.StructLike;
 import org.apache.iceberg.types.Type;
 import org.apache.iceberg.types.Types;
+import org.apache.iceberg.util.DateTimeUtil;
 
 class InternalRecordWrapper implements StructLike {
   private final Function<Object, Object>[] transforms;
